Fix tests and add new docker-composer file.
