+++ query/query0_test.go
@@ -314,7 +314,7 @@ func TestGetUIDInDebugMode(t *testing.T) {
 
 	ctx := context.Background()
 	ctx = context.WithValue(ctx, DebugKey, "true")
-	js, err := processQuery(t, ctx, query)
+	js, err := ProcessQuery(t, ctx, query)
 	require.NoError(t, err)
 	require.JSONEq(t,
 		`{"data": {"me":[{"uid":"0x1","alive":true,"friend":[{"uid":"0x17","name":"Rick Grimes"},{"uid":"0x18","name":"Glenn Rhee"},{"uid":"0x19","name":"Daryl Dixon"},{"uid":"0x1f","name":"Andrea"},{"uid":"0x65"}],"gender":"female","name":"Michonne"}]}}`,
@@ -942,7 +942,7 @@ func TestQueryVarValOrderError(t *testing.T) {
 			}
 		}
 	`
-	_, err := processQuery(t, context.Background(), query)
+	_, err := ProcessQuery(t, context.Background(), query)
 	require.Error(t, err)
 	require.Contains(t, err.Error(), "Cannot sort attribute n of type object.")
 }
@@ -1461,7 +1461,7 @@ func TestDoubleOrder(t *testing.T) {
 		}
 	}
   `
-	_, err := processQuery(t, context.Background(), query)
+	_, err := ProcessQuery(t, context.Background(), query)
 	require.Error(t, err)
 }
 
@@ -1501,7 +1501,7 @@ func TestVarInIneqError(t *testing.T) {
 			}
 		}
   `
-	_, err := processQuery(t, context.Background(), query)
+	_, err := ProcessQuery(t, context.Background(), query)
 	require.Error(t, err)
 }
 
@@ -1729,6 +1729,6 @@ func TestCountUidToVarCombinedWithNormalVar(t *testing.T) {
 }
 
 func TestMain(m *testing.M) {
-	populateCluster()
+	PopulateCluster()
 	os.Exit(m.Run())
 }
